fix: Multiple audio plays in background after fullscreen toggle (#8)
- Previously, the Better Player Controller was being initialized within
the TPStreamPlayer widget's build method. This led to a situation where,
upon widget rebuild, a new controller would be initialized without
properly cleaning up the previous controller. As a result, audio
continued to be sourced from the previous controller.
- This is fixed in this commit by setting the data source on the
controller when it is fetched from the server and initializing the
controller only one-time initState method.